SafeTradeHub-master/requirements.txt
flask
flask-cors
firebase-admin
google-cloud-vision
python-dotenv
requests
beautifulsoup4
selenium
webdriver_manager
